+++ ICD.Connect.Protocol.Network/Ports/Tcp/IcdTcpClient.cs
@@ -4,20 +4,19 @@ using ICD.Common.Properties;
 using ICD.Common.Utils;
 using ICD.Common.Utils.Services.Logging;
 using ICD.Connect.API.Commands;
-using ICD.Connect.API.Nodes;
 using ICD.Connect.Protocol.Network.Settings;
+using ICD.Connect.Protocol.Ports;
 using ICD.Connect.Settings;
 
-namespace ICD.Connect.Protocol.Network.Ports.Udp
+namespace ICD.Connect.Protocol.Network.Ports.Tcp
 {
-	public sealed partial class AsyncUdpClient : AbstractNetworkPort<AsyncUdpClientSettings>
+	public sealed partial class IcdTcpClient : AbstractNetworkPort<IcdTcpClientSettings>
 	{
-		public const ushort DEFAULT_BUFFER_SIZE = 16384;
-		public const string ACCEPT_ALL = "0.0.0.0";
+		public const ushort DEFAULT_PORT = 23;
+		private const ushort DEFAULT_BUFFER_SIZE = 16384;
 
+		private readonly SafeMutex m_SocketMutex;
 		private readonly NetworkProperties m_NetworkProperties;
-
-		private bool m_ListeningRequested;
 
 		#region Properties
 
@@ -27,13 +26,13 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 		public override INetworkProperties NetworkProperties { get { return m_NetworkProperties; } }
 
 		/// <summary>
-		/// Address to accept connections from.
+		/// Get or set the hostname of the remote TCP server.
 		/// </summary>
 		[PublicAPI]
 		public override string Address { get; set; }
 
 		/// <summary>
-		/// Port to accept connections from.
+		/// Get or set the port of the remote TCP server.
 		/// </summary>
 		[PublicAPI]
 		public override ushort Port { get; set; }
@@ -42,22 +41,27 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 		/// Get or set the receive buffer size.
 		/// </summary>
 		[PublicAPI]
-		public int BufferSize { get; set; }
+		public ushort BufferSize { get; set; }
 
 		#endregion
 
+		#region Constructors
+
 		/// <summary>
-		/// Constructor.
+		/// Constructor
 		/// </summary>
-		public AsyncUdpClient()
+		public IcdTcpClient()
 		{
 			m_NetworkProperties = new NetworkProperties();
+			m_SocketMutex = new SafeMutex();
 
+			Port = DEFAULT_PORT;
 			BufferSize = DEFAULT_BUFFER_SIZE;
-			Address = ACCEPT_ALL;
 
 			IcdEnvironment.OnEthernetEvent += IcdEnvironmentOnEthernetEvent;
 		}
+
+		#endregion
 
 		#region Methods
 
@@ -69,31 +73,38 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 			IcdEnvironment.OnEthernetEvent -= IcdEnvironmentOnEthernetEvent;
 
 			base.DisposeFinal(disposing);
+		}
 
-			Disconnect();
+		public void Connect(HostInfo info)
+		{
+			Address = info.Address;
+			Port = info.Port;
+			Connect();
 		}
 
 		/// <summary>
-		/// Sends serial data to a specific endpoint.
+		/// Disconnects from the remote end point
 		/// </summary>
-		/// <param name="data"></param>
-		/// <param name="ipAddress"></param>
-		/// <param name="port"></param>
 		/// <returns></returns>
-		[PublicAPI]
-		public bool SendToAddress(string data, string ipAddress, int port)
+		public override void Disconnect()
 		{
+			if (!m_SocketMutex.WaitForMutex(1000))
+			{
+				Log(eSeverity.Error, "Failed to obtain SocketMutex for disconnect");
+				return;
+			}
+
 			try
 			{
-				if (IsConnected)
-					return SendToAddressFinal(data, ipAddress, port);
-
-				Log(eSeverity.Error, "Unable to send to address - Port is not connected.");
-				return false;
+				DisposeTcpClient();
+			}
+			catch (Exception e)
+			{
+				Log(eSeverity.Error, e, "Failed to disconnect from host {0}:{1}", Address, Port);
 			}
 			finally
 			{
-				UpdateIsConnectedState();
+				m_SocketMutex.ReleaseMutex();
 			}
 		}
 
@@ -102,44 +113,27 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 		#region Private Methods
 
 		/// <summary>
-		/// Handle Ethernet events
+		/// Called when the processor ethernet adapter changes state.
+		/// We connect/disconnect to the endpoint accordingly.
 		/// </summary>
 		/// <param name="adapter"></param>
 		/// <param name="type"></param>
 		private void IcdEnvironmentOnEthernetEvent(IcdEnvironment.eEthernetAdapterType adapter,
 		                                           IcdEnvironment.eEthernetEventType type)
 		{
-#if SIMPLSHARP
-			IcdEnvironment.eEthernetAdapterType adapterType =
-				m_UdpClient == null
-					? IcdEnvironment.eEthernetAdapterType.EthernetUnknownAdapter
-					: IcdEnvironment.GetEthernetAdapterType(m_UdpClient.EthernetAdapterToBindTo);
+			if (m_TcpClient == null)
+				return;
 
-			if (adapter != adapterType && adapterType != IcdEnvironment.eEthernetAdapterType.EthernetUnknownAdapter)
+#if SIMPLSHARP
+			if (adapter != IcdEnvironment.GetEthernetAdapterType(m_TcpClient.EthernetAdapter))
 				return;
 #endif
 			switch (type)
 			{
-				case IcdEnvironment.eEthernetEventType.LinkUp:
-					if (m_ListeningRequested)
-						Connect();
-					break;
-
 				case IcdEnvironment.eEthernetEventType.LinkDown:
-					if (m_UdpClient == null)
-						break;
-#if SIMPLSHARP
-					m_UdpClient.DisableUDPServer();
-#else
-                    m_UdpClient.Dispose();
-#endif
+					Disconnect();
 					break;
-
-				default:
-					throw new ArgumentOutOfRangeException();
 			}
-
-			UpdateIsConnectedState();
 		}
 
 		#endregion
@@ -154,6 +148,8 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 			base.ClearSettingsFinal();
 
 			ApplyConfiguration();
+
+			BufferSize = DEFAULT_BUFFER_SIZE;
 		}
 
 		/// <summary>
@@ -161,7 +157,7 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 		/// </summary>
 		/// <param name="settings"></param>
 		/// <param name="factory"></param>
-		protected override void ApplySettingsFinal(AsyncUdpClientSettings settings, IDeviceFactory factory)
+		protected override void ApplySettingsFinal(IcdTcpClientSettings settings, IDeviceFactory factory)
 		{
 			base.ApplySettingsFinal(settings, factory);
 
@@ -171,20 +167,6 @@ namespace ICD.Connect.Protocol.Network.Ports.Udp
 		#endregion
 
 		#region Console
-
-		/// <summary>
-		/// Calls the delegate for each console status item.
-		/// </summary>
-		/// <param name="addRow"></param>
-		public override void BuildConsoleStatus(AddStatusRowDelegate addRow)
-		{
-			base.BuildConsoleStatus(addRow);
-
-			addRow("Buffer Size", BufferSize);
-#if SIMPLSHARP
-			addRow("Server Status", m_UdpClient == null ? string.Empty : m_UdpClient.ServerStatus.ToString());
-#endif
-		}
 
 		/// <summary>
 		/// Gets the child console commands.