+++ ICD.Connect.Protocol.Network/Ports/Udp/IcdUdpClient.SimplSharp.cs
@@ -10,7 +10,7 @@ using ICD.Connect.Protocol.Ports;
 
 namespace ICD.Connect.Protocol.Network.Ports.Udp
 {
-	public sealed partial class AsyncUdpClient
+	public sealed partial class IcdUdpClient
 	{
 		[CanBeNull]
 		private UDPServer m_UdpClient;